Refactor certificate_managed changes checks
to be able to use them in a wrapper module
